pyproject.toml
[tool.poetry]
name = "hedwigbackend"
version = "0.1.0"
description = ""
authors = ["inerv <voldemar.krs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.1"
djangorestframework = "^3.9"
markdown = "^3.0"
django-filter = "^2.0"
django-environ = "^0.4.5"
Celery = "^4.2"
celery_once = "^2.0"
django-redis = "^4.10"
django-cors-headers = "^2.4"
django-behaviors = "^0.4.1"
drf-nested-routers = "^0.91.0"
redis = "^3.0"
psycopg2 = "^2.7"
psycopg2-binary = "^2.7"
djangorestframework_simplejwt = "^3.3"

[tool.poetry.dev-dependencies]
pytest = "^4.0"
pytest-django = "^3.4"
mixer = "^6.1"
isort = "^4.3"
flake8 = "^3.6"
flake8-isort = "^2.6"
flake8-bugbear = "^18.8"
flake8-polyfill = "^1.0"
flake8-print = "^3.1"
coverage = "^4.5"
pytest-cov = "^2.6"
pytest-env = "^0.6.2"
pytest-forked = "^0.2.0"
pytest-freezegun = "^0.3.0"
pytest-mock = "^1.10"
pytest-xdist = "^1.25"
flake8-builtins = "^1.4"
flake8-type-annotations = "^0.1.0"
flake8-django = "^0.0.3"
flake8-pytest = "^1.3"
flake8-mock = "^0.3.0"
nplusone = "^1.0"
django-querycount = "^0.7.0"
django-debug-toolbar = "^1.11"
simplejson = "^3.16"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
